+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 
 ## [Unreleased]
+### Added
+- Account can now logout
 ## [v0.5.0] - 2021-10-06
 ### Added
 - More debug logging messages